refactor(eval-core): 消除 sealed plan 依赖环
