+++ src/eval-core/compiler/index.ts
@@ -56,11 +56,11 @@ import {
   validateAnalysisInputs,
   validateDefinitionSemantics,
 } from './validation.js';
-import { sealRunPlan } from '../internal/sealed-run-plan.js';
+import { sealRunPlan } from '../contracts/sealed-run-plan.js';
 
 export * from './errors.js';
 export * from './types.js';
-export { assertSealedRunPlan } from '../internal/sealed-run-plan.js';
+export { assertSealedRunPlan } from '../contracts/sealed-run-plan.js';
 export { validateDefinitionSemantics } from './validation.js';
 
 interface StageExtensions {
