+++ src/eval-core/contracts/sealed-run-plan.ts
@@ -1,6 +1,6 @@
-import type { JsonValue } from '../contracts/json.js';
-import { deepFreezeCanonicalJson } from '../contracts/json.js';
-import type { RunPlan } from '../contracts/plans.js';
+import type { JsonValue } from './json.js';
+import { deepFreezeCanonicalJson } from './json.js';
+import type { RunPlan } from './plans.js';
 
 export type DeepReadonly<T> = T extends JsonValue
   ? T extends readonly (infer Item)[]
